+++ hooks/contact_status_changed_test.go
@@ -14,46 +14,45 @@ func TestContactStatusChanged(t *testing.T) {
 	db := testsuite.DB()
 
 	// make sure cathyID contact is active
-	db.Exec(`UPDATE contacts_contact SET is_blocked = FALSE WHERE id = $1`, models.CathyID)
-	db.Exec(`UPDATE contacts_contact SET is_stopped = FALSE WHERE id = $1`, models.CathyID)
+	db.Exec(`UPDATE contacts_contact SET status = 'A', is_blocked = FALSE, is_stopped = FALSE WHERE id = $1`, models.CathyID)
 
 	tcs := []HookTestCase{
-		HookTestCase{
+		{
 			Modifiers: ContactModifierMap{
 				models.CathyID: []flows.Modifier{modifiers.NewStatus(flows.ContactStatusBlocked)},
 			},
 			SQLAssertions: []SQLAssertion{
-				SQLAssertion{
-					SQL:   `select count(*) from contacts_contact where id = $1 AND is_blocked = TRUE`,
+				{
+					SQL:   `select count(*) from contacts_contact where id = $1 AND status = 'B' AND is_blocked = TRUE`,
 					Args:  []interface{}{models.CathyID},
 					Count: 1,
 				},
 			},
 		},
-		HookTestCase{
+		{
 			Modifiers: ContactModifierMap{
 				models.CathyID: []flows.Modifier{modifiers.NewStatus(flows.ContactStatusStopped)},
 			},
 			SQLAssertions: []SQLAssertion{
-				SQLAssertion{
-					SQL:   `select count(*) from contacts_contact where id = $1 AND is_stopped = TRUE`,
+				{
+					SQL:   `select count(*) from contacts_contact where id = $1 AND status = 'S' AND is_stopped = TRUE`,
 					Args:  []interface{}{models.CathyID},
 					Count: 1,
 				},
 			},
 		},
-		HookTestCase{
+		{
 			Modifiers: ContactModifierMap{
 				models.CathyID: []flows.Modifier{modifiers.NewStatus(flows.ContactStatusActive)},
 			},
 			SQLAssertions: []SQLAssertion{
-				SQLAssertion{
-					SQL:   `select count(*) from contacts_contact where id = $1 AND is_stopped = FALSE`,
+				{
+					SQL:   `select count(*) from contacts_contact where id = $1 AND status = 'A' AND is_stopped = FALSE`,
 					Args:  []interface{}{models.CathyID},
 					Count: 1,
 				},
-				SQLAssertion{
-					SQL:   `select count(*) from contacts_contact where id = $1 AND is_blocked = FALSE`,
+				{
+					SQL:   `select count(*) from contacts_contact where id = $1 AND status = 'A' AND is_blocked = FALSE`,
 					Args:  []interface{}{models.CathyID},
 					Count: 1,
 				},