+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+v5.7.12
+----------
+ * Fix inserting channel logs and add test
+
+v5.7.11
+----------
+ * Always write contact.status when writing is_blocked or is_stopped
+ * Convert IVR code to use goflow's httpx package
+
 v5.7.10
 ----------
  * Tweak goreleaser config to include subdirectories inside docs folder